requirements.txt
celery==4.4.2
fabric3==1.14.post1